Cleanup the RQ hooks.
