+++ apps/studio/data/api-keys/keys.ts
@@ -1,8 +1,5 @@
 export const apiKeysKeys = {
   list: (projectRef?: string) => ['projects', projectRef, 'api-keys'] as const,
   single: (projectRef?: string, id?: string) => ['projects', projectRef, 'api-keys', id] as const,
-}
-
-export const legacyAPIKeysEnabledKeys = {
-  enabled: (projectRef?: string) => ['projects', projectRef, 'api-keys', 'legacy'] as const,
+  status: (projectRef?: string) => ['projects', projectRef, 'api-keys', 'legacy'] as const,
 }
